Fix restake amount calculation and remove unused import
